move HomePage to WarehouseList component
