+++ src/components/WarehouseList/WarehouseList.scss
@@ -1,0 +1,118 @@
+@use "../../styles/partials/variables" as *;
+@use "../../styles/partials/mixins" as *;
+@use "../../styles/partials/fonts" as *;
+
+// Top of the table with the sorting buttons
+.sorty-selectors {
+	display: none;
+	@media (min-width: $tablet-breakpoint) {
+		display: flex;
+	}
+	padding: 0 1.25rem;
+	background-color: $slate;
+	justify-content: space-around;
+	align-items: center;
+
+	&__item {
+		display: flex;
+		justify-content: flex-start;
+		width: 7rem;
+		&:nth-child(4) {
+			width: 10rem;
+		}
+		& img {
+			width: 25px;
+		}
+		@media (min-width: $pc-breakpoint) {
+			width: 10.75rem;
+		}
+	}
+}
+
+// Make the Actions column small
+@mixin col5 {
+	display: flex;
+	justify-content: space-between;
+	width: 100%;
+
+	@media (min-width: $tablet-breakpoint) {
+		width: 3rem;
+		justify-content: flex-start;
+	}
+
+	& img {
+		width: 25px;
+	}
+}
+
+// For the Actions header
+.col5 {
+	@include col5;
+}
+
+.warehouse-container {
+	display: flex;
+	justify-content: flex-start;
+	align-items: center;
+	flex-wrap: wrap;
+	padding: 0 1.25rem;
+	@media (min-width: $tablet-breakpoint) {
+		justify-content: space-around;
+	}
+
+	&__item {
+		width: 50%;
+		margin-top: 1rem;
+		word-wrap: break-word;
+		// For the divs with the Actions buttons
+		&--col5 {
+			@include col5;
+		}
+
+		@media (min-width: $tablet-breakpoint) {
+			width: 7rem;
+			&:nth-child(4) {
+				width: 10rem;
+			}
+		}
+
+		@media (min-width: $pc-breakpoint) {
+			width: 10.75rem;
+		}
+	}
+
+	& span {
+		display: flex;
+	}
+
+	& h3,
+	h4,
+	p {
+		margin: 0;
+		margin-top: 0.25rem;
+	}
+}
+
+.divider {
+	width: 100%;
+	height: 1px;
+	background-color: $cloud;
+	margin-top: 1rem;
+}
+
+.search {
+	@include interactive-components;
+	background-image: url("../../assets/Icons/search-24px.svg");
+	background-repeat: no-repeat;
+	background-position: right 1rem center;
+	width: calc(100% - 2rem);
+	margin-bottom: 16px;
+}
+
+.btn__add {
+	@include btn;
+	width: 100%;
+	background-color: $instock-indigo;
+	color: $white;
+	margin-bottom: 16px;
+}
